+++ recommendation-service/src/main/java/com/gokdenizozkan/ddd/recommendationservice/entity/Rating.java
@@ -1,18 +1,18 @@
 package com.gokdenizozkan.ddd.recommendationservice.entity;
 
 public enum Rating {
-    EXCELLENT(5),
-    SATISFACTORY(4),
-    DECENT(3),
-    LACKING(2),
-    POOR(1);
+    EXCELLENT(5F),
+    SATISFACTORY(4F),
+    DECENT(3F),
+    LACKING(2F),
+    POOR(1F);
 
-    public final int value;
-    public static final Integer MAX_RATING_VALUE = 5;
-    public static final Integer MIN_RATING_VALUE = 1;
+    public final Float value;
+    public static final Float MAX_RATING_VALUE = 5F;
+    public static final Float MIN_RATING_VALUE = 1F;
     private final int availableRatingCount;
 
-    Rating(int value) {
+    Rating(Float value) {
         this.value = value;
         this.availableRatingCount = 5;
     }
